Fix GUI import shadowing and finalize Wayland/AUR support
- Resolve ImportError by moving src/pnp/gui.py to src/pnp/gui/__init__.py
- Implement GTK4-compatible system tray using pure DBus SNI protocol
- Add Wayland/X11 detection and scaling overrides in gui/__init__.py
- Update pnp.service with Wayland environment and %t specifier
- Modernize PKGBUILD for AUR with rye and python-installer
- Update README with compatibility and packaging documentation
